requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.21.41
botocore==1.24.41
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-ckeditor==6.2.0
django-environ==0.8.1
django-extensions==3.1.5
django-js-asset==2.0.0
django-parler==2.3
django-ranged-response==0.2.0
django-rosetta==0.9.8
django-simple-captcha==0.5.17
django-storages==1.12.3
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
Pillow==9.1.0
polib==1.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.9
Werkzeug==2.0.2
whitenoise==5.3.0
